Create ajax-request-model.py
In this example, we're using the requests.get() method to make a GET request to the URL https://example.com/api/data. We're then checking the status code of the response using response.status_code. If the status code is 200 (which means the request was successful), we print the content of the response using response.content. Otherwise, we print an error message.

You can also include data or headers in your Ajax request by passing them as parameters to the requests.get() method.