Create simple-scrape-model.py
To scrape DOM HTML in Python, you can use the BeautifulSoup library. 
In this example, we're making a GET request to the URL https://example.com using the requests.get() method. We're then using BeautifulSoup to parse the HTML content of the response using the html.parser parser.

We can then use soup.find_all('a') to find all the links on the page. This returns a list of Tag objects, which we can iterate over to print the href attribute of each link using link.get('href').

You can also use other methods provided by BeautifulSoup to extract specific elements from the HTML, such as soup.find() to find the first occurrence of an element, or soup.select() to find elements based on CSS selectors.